pyproject.toml
[tool.poetry]
name = "dip1-1"
version = "0.1.0"
description = ""
authors = ["sacred70 <sacred70@mail.ru>"]


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
pydantic = ">=1.10,<2.0"
requests = "^2.31.0"
django-request-logging = "^0.7.5"
pytest = "^7.4.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
freezegun = "^1.2.2"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
